+++ pf_business/src/main/java/com/nnte/pf_business/mapper/workdb/functions/PlateformFunctions.java
@@ -7,7 +7,7 @@ import java.util.Date;
  * 自动代码 请勿更改 <2020-06-21 13:55:29>
  */
 public class PlateformFunctions extends BaseModel {
-    @DBPKColum private Long id;
+    @DBPKColum private Integer id;
     private String menuCode;
     private String funCode;
     private String funName;
@@ -18,8 +18,8 @@ public class PlateformFunctions extends BaseModel {
 
     public PlateformFunctions(){}
 
-    public Long  getId(){ return id;}
-    public void setId(Long  id){ this.id = id;}
+    public Integer  getId(){ return id;}
+    public void setId(Integer  id){ this.id = id;}
     public String  getMenuCode(){ return menuCode;}
     public void setMenuCode(String  menuCode){ this.menuCode = menuCode;}
     public String  getFunCode(){ return funCode;}
